+++ src/Components/BestProducts.jsx
@@ -1,0 +1,17 @@
+import ItemCard from './ItemCard';
+
+function BestProducts() {
+  return (
+    <>
+      <div>베스트상품</div>
+      <div>
+        <ItemCard />
+        <ItemCard />
+        <ItemCard />
+        <ItemCard />
+      </div>
+    </>
+  );
+}
+
+export default BestProducts;
