+++ src/Components/ItemCard.css
@@ -1,0 +1,7 @@
+* {
+  box-sizing: border-box;
+}
+
+html {
+  font-size: 62.5%;
+}
