+++ src/Components/ItemCard.jsx
@@ -1,0 +1,13 @@
+function ItemCard() {
+  return (
+    <div class="CardDiv">
+      <img></img>
+      <div>아이패드 미니 팝니다.</div>
+      <div>500,000원</div>
+      <button />
+      <div>1000</div>
+    </div>
+  );
+}
+
+export default ItemCard;
