+++ src/Components/PaginationBar.jsx
@@ -1,0 +1,15 @@
+function PaginationBar() {
+  return (
+    <div>
+      <button>&lt;</button>
+      <button>1</button>
+      <button>2</button>
+      <button>3</button>
+      <button>4</button>
+      <button>5</button>
+      <button>&gt;</button>
+    </div>
+  );
+}
+
+export default PaginationBar;
