+++ src/Components/TopNavigation.jsx
@@ -1,0 +1,9 @@
+function TopNavigation() {
+  return (
+    <div>
+      <button /> <button>자유게시판</button> <button>중고마켓</button> <button>로그인</button>
+    </div>
+  );
+}
+
+export default TopNavigation;
