--- src/Items.jsx
@@ -1,136 +1,0 @@
-function Items() {
-  return (
-    <>
-      <div>
-        <button /> <button>자유게시판</button> <button>중고마켓</button> <button>로그인</button>
-      </div>
-      <div>
-        <div>베스트상품</div>
-        <div>
-          <div>
-            <img></img>
-            <div>아이패드 미니 팝니다.</div>
-            <div>500,000원</div>
-            <button />
-            <div>1000</div>
-          </div>
-          <div>
-            <img></img>
-            <div>아이패드 미니 팝니다.</div>
-            <div>500,000원</div>
-            <button />
-            <div>1000</div>
-          </div>
-          <div>
-            <img></img>
-            <div>아이패드 미니 팝니다.</div>
-            <div>500,000원</div>
-            <button />
-            <div>1000</div>
-          </div>
-          <div>
-            <img></img>
-            <div>아이패드 미니 팝니다.</div>
-            <div>500,000원</div>
-            <button />
-            <div>1000</div>
-          </div>
-        </div>
-        <div>
-          <div>판매중인 상품</div>
-          <form>
-            <input placeholder="검색할 상품을 입력해주세요"></input>
-            <button>상품 등록하기</button>
-            <select>
-              <option value="option1">최신순</option>
-              <option value="option2">좋아요순</option>
-            </select>
-          </form>
-          <div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-            <div>
-              <img></img>
-              <div>아이패드 미니 팝니다.</div>
-              <div>500,000원</div>
-              <button />
-              <div>1000</div>
-            </div>
-          </div>
-        </div>
-        <div>
-          <button>&lt;</button>
-          <button>1</button>
-          <button>2</button>
-          <button>3</button>
-          <button>4</button>
-          <button>5</button>
-          <button>&gt;</button>
-        </div>
-      </div>
-    </>
-  );
-}
-
-export default Items;
